# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Ocorrências",
        "description": "API para gerenciamento de ocorrências, despachos e comandos. Suporta integrações externas e sistemas internos.",
        "contact": {
            "email": "suporte@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8089",
            "description": "Servidor de API"
        }
    ],
    "paths": {
...
        "/api/occurrences/{occurrenceId}/dispatches": {
            "post": {
                "tags": [
                    "Dispatches"
                ],
                "summary": "Criar um novo despacho",
                "description": "Cria um novo despacho para uma ocorrência específica. Retorna um comando que pode ser consultado posteriormente para verificar o status",
                "operationId": "createDispatch",
                "parameters": [
                    {
                        "name": "occurrenceId",
                        "in": "path",
                        "description": "ID único da ocorrência (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDispatchRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Comando de criação de despacho aceito para processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Command"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ocorrência não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação - Dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado - API Key inválida ou ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/dispatches/{dispatchId}/close": {
            "post": {
                "tags": [
                    "Dispatches"
                ],
                "summary": "Fechar um despacho",
                "description": "Fecha um despacho específico. Retorna um comando que pode ser consultado posteriormente para verificar o status",
                "operationId": "closeDispatch",
                "parameters": [
                    {
                        "name": "dispatchId",
                        "in": "path",
                        "description": "ID único do despacho (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440001"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Comando de fechamento de despacho aceito para processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Command"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Despacho não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado - API Key inválida ou ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/dispatches/{dispatchId}/status": {
            "patch": {
                "tags": [
                    "Dispatches"
                ],
                "summary": "Atualizar status de um despacho",
                "description": "Atualiza o status de um despacho. Retorna um comando que pode ser consultado posteriormente para verificar o status",
                "operationId": "updateDispatchStatus",
                "parameters": [
                    {
                        "name": "dispatchId",
                        "in": "path",
                        "description": "ID único do despacho (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440001"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDispatchStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Comando de atualização de status aceito para processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Command"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Despacho não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação - Status inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Requisição duplicada - Idempotency Key já utilizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado - API Key inválida ou ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "idempotencyKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Command": {
                "title": "Comando",
                "description": "Modelo de dados de um comando assíncrono",
                "properties": {
                    "command_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440002"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "completed",
                            "failed"
                        ],
                        "example": "completed"
                    },
                    "result": {
                        "description": "Resultado do comando quando status é 'completed'",
                        "type": "object",
                        "nullable": true
                    },
                    "error": {
                        "description": "Mensagem de erro quando status é 'failed'",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Error": {
                "title": "Erro",
                "description": "Modelo de resposta de erro",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid"
                    },
                    "errors": {
                        "description": "Detalhes dos erros de validação",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "CreateDispatchRequest": {
                "title": "Criar Despacho",
                "description": "Dados necessários para criar um novo despacho",
                "required": [
                    "resourceCode"
                ],
                "properties": {
                    "resourceCode": {
                        "description": "Código do recurso no formato XX-YY ou XXX-YY (ex: ABT-12, UR-05)",
                        "type": "string",
                        "pattern": "^[A-Z]{2,3}-\\d{2}$",
                        "example": "ABT-12"
                    }
                },
                "type": "object"
            },
            "UpdateDispatchStatusRequest": {
                "title": "Atualizar Status do Despacho",
                "description": "Dados necessários para atualizar o status de um despacho",
                "required": [
                    "statusCode"
                ],
                "properties": {
                    "statusCode": {
                        "description": "Novo status do despacho",
                        "type": "string",
                        "enum": [
                            "assigned",
                            "en_route",
                            "on_site",
                            "closed"
                        ],
                        "example": "on_site"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Chave de API para autenticacao.",
                "name": "X-API-Key",
                "in": "header"
            },
            "idempotencyKey": {
                "type": "apiKey",
                "description": "Chave de idempotencia para operacoes de escrita.",
                "name": "Idempotency-Key",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "apiKey": []
        }
    ]
}